refactor: remove etherscan link on token select
